Moved renderer out of the src directory
