+++ tsconfig.json
@@ -1,11 +1,14 @@
 {
   "extends": "@electron-toolkit/tsconfig/tsconfig.json",
   "include": [
-    "src/renderer/src/env.d.ts",
-    "src/renderer/src/**/*",
-    "src/renderer/src/**/*.svelte",
-    "src/preload/*.d.ts"
-  ],
+    // "src/renderer/src/env.d.ts",
+    // "src/renderer/src/**/*",
+    // "src/renderer/src/**/*.svelte",
+    "renderer/env.d.ts",
+    "renderer/**/*",
+    "renderer/**/*.svelte",
+    "preload/*.d.ts"
+, "renderer/lib", "renderer/state", "renderer/main.ts"  ],
   "compilerOptions": {
     "verbatimModuleSyntax": true,
     "useDefineForClassFields": true,
